requirements.txt
pandas
ruff
pytest
assertpy
fastapi
uvicorn
PyJWT
kubernetes
python-dotenv
cryptography